pyproject.toml
[project]
name = "binance-testnet-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "loguru>=0.7.3",
    "pydantic>=2.12.4",
    "pydantic-settings>=2.12.0",
    "python-binance>=1.0.32",
    "typer>=0.20.0",
]
